Restructuring with accordance to standards
