+++ .settings/org.eclipse.core.resources.prefs
@@ -1,2 +1,2 @@
 eclipse.preferences.version=1
-encoding//Practice/LeetCode/Problems/Solution1672.java=UTF-8
+encoding//src/main/java/com/problems/leetcode/Solution1672.java=UTF-8
