--- Practice/LeetCode/Problems/Solution1281.java
@@ -1,39 +1,0 @@
-/*
-1281. Subtract the Product and Sum of Digits of an Integer
-https://leetcode.com/problems/subtract-the-product-and-sum-of-digits-of-an-integer/
-
-Given an integer number n, return the difference between the product of its digits and the sum of its digits.
- 
-
-Example 1:
-
-Input: n = 234
-Output: 15 
-Explanation: 
-Product of digits = 2 * 3 * 4 = 24 
-Sum of digits = 2 + 3 + 4 = 9 
-Result = 24 - 9 = 15
-Example 2:
-
-Input: n = 4421
-Output: 21
-Explanation: 
-Product of digits = 4 * 4 * 2 * 1 = 32 
-Sum of digits = 4 + 4 + 2 + 1 = 11 
-Result = 32 - 11 = 21
-
- */
-
-class Solution1281 {
-	public int subtractProductAndSum(int n) {
-		int mul = 1;
-		int sum = 0;
-		while (n > 0) {
-			int k = n % 10;
-			n = n / 10;
-			mul = mul * k;
-			sum = sum + k;
-		}
-		return mul - sum;
-	}
-}
